# Fast implementation of maximum cardinality search ordering algorithm.