I have a library [CliqueTrees.jl](https://github.com/AlgebraicJulia/CliqueTrees.jl/tree/main) with a fast implementation of the maximum cardinality algorithm. Here are some benchmarks.

```julia
using BenchmarkTools, MatrixMarket, PowerModels, SuiteSparseMatrixCollection

ssmc = ssmc_db()
name = "bcsstk17"
matrix = mmread(joinpath(fetch_ssmc(ssmc[ssmc.name .== name, :], format="MM")[1], "$(name).mtx"))
```

**current implementation**

```julia
julia> @benchmark PowerModels._mcs($matrix)
BenchmarkTools.Trial: 7 samples with 1 evaluation per sample.
 Range (min … max):  778.347 ms … 849.834 ms  ┊ GC (min … max): 4.17% … 6.36%
 Time  (median):     801.612 ms               ┊ GC (median):    5.25%
 Time  (mean ± σ):   803.112 ms ±  23.522 ms  ┊ GC (mean ± σ):  5.17% ± 0.72%

  █   █       █      █ █     █                                █  
  █▁▁▁█▁▁▁▁▁▁▁█▁▁▁▁▁▁█▁█▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  778 ms           Histogram: frequency by time          850 ms <

 Memory estimate: 1.50 GiB, allocs estimate: 238040.
 ```
 
 **new implementation**
 
```julia
julia> @benchmark PowerModels._mcs($matrix)
BenchmarkTools.Trial: 4301 samples with 1 evaluation per sample.
 Range (min … max):  1.083 ms …   5.601 ms  ┊ GC (min … max): 0.00% … 75.50%
 Time  (median):     1.128 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.160 ms ± 153.791 μs  ┊ GC (mean ± σ):  2.05% ±  6.05%

  ▆▇▆█▇▅▅▄▃▃▂▂      ▁▁▁▁                                      ▂
  ████████████████▇▇█████▇█▇▆▇▅▆▇▅▄▅▇▃▆▃▃▃▃▆▃▁▁▃▅▅▃▄▃▄▁▃▃▅▁▁▃ █
  1.08 ms      Histogram: log(frequency) by time      1.78 ms <

 Memory estimate: 514.88 KiB, allocs estimate: 18.
```

The new implementation is linear time, and the current one is quadratic (cubic?), so the difference will get worse with larger matrices. You can take a look at the new implementation [here](https://github.com/AlgebraicJulia/CliqueTrees.jl/blob/b7b64bcc3afb60b5aadfa0cc1ba8b0b145b3443b/src/elimination_algorithms.jl#L522).