@odow 2025-02-24
How long does a very large example take? I think we'd be hesitant to adopt a new direct dependency just to solve a second, even though I'll admit it looks like you have done a nice job :smile: