@samuelsonric 2025-02-24
74752x74532 with 596992 stored entries

```julia
name = "bcsstk17"
```

**current implementation**

```julia
julia> @benchmark PowerModels._mcs($matrix)
BenchmarkTools.Trial: 1 sample with 1 evaluation per sample.
 Single result which took 43.177 s (8.29% GC) to evaluate,
 with a memory estimate of 71.03 GiB, over 1563700 allocations.
```

**new implementation**
julia> @benchmark PowerModels._mcs($matrix)
BenchmarkTools.Trial: 1964 samples with 1 evaluation per sample.
 Range (min … max):  2.163 ms …  22.294 ms  ┊ GC (min … max): 0.00% … 84.34%
 Time  (median):     2.381 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.541 ms ± 624.763 μs  ┊ GC (mean ± σ):  6.08% ±  9.25%

      ▂█▅▂                                                     
  ▂▃▃▄████▇▄▄▃▂▂▂▂▂▂▂▂▂▃▃▄▄▄▃▃▃▂▂▂▂▂▂▁▂▂▁▁▂▂▂▁▂▂▂▁▁▂▁▂▁▁▂▂▂▂▂ ▃
  2.16 ms         Histogram: frequency by time        4.07 ms <

 Memory estimate: 3.42 MiB, allocs estimate: 18.
 ```