@odow 2025-02-24
Sure, that's a nice example, but what about one that occurs in the types of models that PowerModels solves?