@samuelsonric 2025-02-24
> Sure, that's a nice example, but what about one that occurs in the types of models that PowerModels solves?

Oh! I see what you mean --- I'll run some tests.

CliqueTrees is a package for constructing chordal extensions; after this PR, I will try optimizing the functions [_chordal_extension](https://github.com/lanl-ansi/PowerModels.jl/blob/8824212b3d2f35f94f7dd72f37ff76e3419c478f/src/form/wrm.jl#L328) and [_maximal_cliques](https://github.com/lanl-ansi/PowerModels.jl/blob/8824212b3d2f35f94f7dd72f37ff76e3419c478f/src/form/wrm.jl#L352).