@odow 2025-02-24
This isn't my subject area, so I'll defer to @ccoffrin, but it looks like this code was added in https://github.com/lanl-ansi/PowerModels.jl/pull/333, and I couldn't find any related issues inn the last 7 years, so it may be that the current implementation is sufficient.

I assume most matrix sizes are <10^3. Larger and we get to the point where we can't solve the underlying optimization model so the ability to build the problem becomes moot.

That's not to say that there isn't room for optimizing the implementations, but the bar for adding a new dependency is very high.