@ccoffrin 2025-02-28
I do not want to add an external package dependency for improved performance on this, but I am open to performance improvements that only use the current set of dependencies.

This code is only used in one very specific model, an SPD relaxation taking advantage of matrix sparsity. I strongly suspect that the SDP solver runtime will dwarf the model build time for these models.