@samuelsonric 2025-02-28
> I do not want to add an external package dependency for improved performance on this, but I am open to performance improvements that only use the current set of dependencies.
> 
> This code is only used in one very specific model, an SPD relaxation taking advantage of matrix sparsity. I strongly suspect that the SDP solver runtime will dwarf the model build time for these models.

I tried benchmarking the `instantiate_model` function a on the TYP models in [this](https://github.com/power-grid-lib/pglib-opf/blob/master/BASELINE.md) (quite helpful) data set, and by the time I got halfway down the list, instantiation was taking several minutes; the `_mcs` call is a tiny part of that.

In any case, it's just one function, so I am happy to copy in the code.