@odow 2025-03-04
I don't think this is what @ccoffrin had in mind. This would add a few hundred lines of untested code that is pretty tangential to the core of PowerModels. Yes, we could add the CliqueTrees tests, but that's a bit beside the point.

> and by the time I got halfway down the list, instantiation was taking several minutes; the _mcs call is a tiny part of that.

Is this an argument that we don't need this PR? There are other parts of the code that are much slower.