@samuelsonric 2025-03-04
> Is this an argument that we don't need this PR? There are other parts of the code that are much slower.

Yeah, it doesn't seem to make much of a difference.